+++ src/main/java/com/solera/rubiogoncalves/airlinesapp/examples/GreetingController.java
@@ -4,7 +4,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-
+/**
+ * So far, our HomeController is simple and has no dependencies.
+ * We could make it more realistic by introducing an extra component
+ * to store the greeting (perhaps in a new controller).
+ *
+ * The following example shows how to do so:
+ */
 @Controller
 public class GreetingController {
