--- src/test/java/com/solera/rubiogoncalves/airlinesapp/webTestExamples/HttpRequestTest.java
@@ -1,30 +1,0 @@
-package com.solera.rubiogoncalves.airlinesapp.webTestExamples;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.test.web.server.LocalServerPort;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class HttpRequestTest {
-
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Test
-    public void greetingShouldReturnDefaultMessage() throws Exception {
-        assertThat(
-                this.restTemplate.getForObject(
-                        "http://localhost:" + port + "/",
-                        String.class
-                )
-        ).contains("Hello, World");
-    }
-}
